+++ climpred/classes.py
@@ -4,24 +4,31 @@ from .bootstrap import (
     bootstrap_perfect_model,
     bootstrap_uninit_pm_ensemble_from_control,
 )
+from .checks import (
+    has_dims,
+    is_initialized,
+    is_xarray,
+    match_initialized_dims,
+    match_initialized_vars,
+)
 from .prediction import (
     compute_hindcast,
     compute_perfect_model,
     compute_persistence,
     compute_uninitialized,
 )
-from .checks import (
-    has_dims,
-    is_xarray,
-    is_initialized,
-    match_initialized_dims,
-    match_initialized_vars,
+from .smoothing import (
+    smooth_goddard_2013,
+    spatial_smoothing_xesmf,
+    spatial_smoothing_xrcoarsen,
+    temporal_smoothing,
 )
-
 
 # ----------
 # Aesthetics
 # ----------
+
+
 def _display_metadata(self):
     """
     This is called in the following case:
@@ -90,6 +97,80 @@ class PredictionEnsemble:
     # https://stackoverflow.com/questions/1535327/how-to-print-objects-of-class-using-print
     def __repr__(self):
         return _display_metadata(self)
+
+    def smooth(self, smooth_kws='goddard2013'):
+        """Smooth all entries of PredictionEnsemble in the same manner to be
+        able to still calculate prediction skill afterwards.
+
+        Args:
+          xobj (xarray object):
+            decadal prediction ensemble output.
+
+        Attributes:
+            smooth_kws (dict or str): Dictionary to specify the dims to
+                smooth compatible with `spatial_smoothing_xesmf`,
+                `temporal_smoothing` or `spatial_smoothing_xrcoarsen`.
+                Shortcut for Goddard et al. 2013 recommendations:
+                'goddard2013'
+
+        Example:
+        >>> PredictionEnsemble.smooth(smooth_kws={'time': 2,
+            'lat': 5, 'lon': 4'})
+        >>> PredictionEnsemble.smooth(smooth_kws='goddard2013')
+        """
+        # get proper smoothing function based on smooth args
+        if isinstance(smooth_kws, str):
+            if 'goddard' in smooth_kws:
+                smooth_fct = smooth_goddard_2013
+                smooth_kws = {'lead': 4}  # default
+            else:
+                raise ValueError(
+                    'Please provide from list of available smoothings: \
+                     ["goddard2013"]'
+                )
+        elif isinstance(smooth_kws, dict):
+            non_time_dims = [
+                dim for dim in smooth_kws.keys() if dim not in ['time', 'lead']
+            ]
+            if len(non_time_dims) > 0:
+                non_time_dims = non_time_dims[0]
+            print(non_time_dims, 'non_time_dims')
+            # goddard when time_dim and lon/lat given
+            if ('lon' in smooth_kws or 'lat' in smooth_kws) and (
+                'lead' in smooth_kws or 'time' in smooth_kws
+            ):
+                smooth_fct = smooth_goddard_2013
+            # fail goddard and fall back to xrcoarsen when
+            # coarsen dim and time_dim provided
+            elif (
+                (non_time_dims is not [])
+                and (non_time_dims in list(self.initialized.dims))
+                and ('lead' in smooth_kws or 'time' in smooth_kws)
+            ):
+                smooth_fct = smooth_goddard_2013
+            # else only one smoothing operation
+            elif 'lon' in smooth_kws or 'lat' in smooth_kws:
+                smooth_fct = spatial_smoothing_xesmf
+            elif 'lead' in smooth_kws:
+                smooth_fct = temporal_smoothing
+            elif non_time_dims in list(self.initialized.dims):
+                smooth_fct = spatial_smoothing_xrcoarsen
+            else:
+                raise ValueError(
+                    'Please provide kwargs to fulfill functions: \
+                     ["spatial_smoothing_xesmf", "temporal_smoothing", \
+                     "spatial_smoothing_xrcoarsen"].'
+                )
+        else:
+            raise ValueError(
+                'Please provide kwargs as str or dict and not', type(smooth_kws)
+            )
+        self.initialized = smooth_fct(self.initialized, smooth_kws)
+        # check for other objects in  PredictionEnsemble
+        # better: for obj in self: do smooth
+        if isinstance(self.uninitialized, xr.Dataset):
+            self.uninitialized = smooth_fct(self.uninitialized, smooth_kws)
+        # self.reference not implemented
 
 
 class PerfectModelEnsemble(PredictionEnsemble):