+++ CHANGELOG.rst
@@ -2,15 +2,15 @@
 Changelog History
 =================
 
-climpred v1.1 (2019-08-##)
+climpred v1.1 (2019-09-##)
 ============================
 
 Features
 --------
-- Write information about skill computation to netcdf attributes (:pr:`213`) `Aaron Spring`_
+- Write information about skill computation to netcdf attributes(:pr:`213`) `Aaron Spring`_
+- Temporal and spatial smoothing module (:pr:`224`) `Aaron Spring`_
 - Add metric `brier_score`, `threshold_brier_score`, `crpss_es` and `less` (:pr:`232`) `Aaron Spring`_
 - allow `compute_hindcast` and `compute_perfect_model` to specify which dimension `dim` to calculate metric over (:pr:`232`) `Aaron Spring`_
-
 
 Bug Fixes
 ---------